+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 2021-11-27
+- Improvement(Leaper): Remove unnecessary polling the mutex inside fn try_catch(|result| {...}).
+- Explicit cancelation support
+- Add more examples
+- A few optimizations.
+
 ## [0.1.5] - 2021-11-27
 - Remove unnecessary async topping.
 - Two more Functionalities added (fn is_flowing and fn is_leaping) to the handle.